Add files via upload
/*
 * TECHNICAL DATA OF THE - Vatio (VATT):
 * Total Supply   : 100,000,000
 * Name             : Vatio
 * Symbol           : VATT
 * Decimals         : 9
 */

/*
Vatio (VATT)
The currency of the new energy age”, “Decentralized power”
* Purpose:
Vatio is the native token of our ecosystem. It is designed to accrue value 
through the execution, development, and management of profitable and 
self-sustainable projects.

* Value Generation:
- Liquidity is generated via the automatic liquidity mechanism on each transaction.
- Revenues from real-world projects are used to support liquidity and burning.
- Token utility is linked to ecosystem services and governance.

This approach ensures VATT maintains a deflationary model with increasing utility and value.
"We are building the future of digital energy together with you. Thank you for believing in the Vatio VATT project."
*/